added a new protect subcommand that sets a simpledb attribute that is now used to prevent stack destruction

+++ lib/simple_deploy/stack.rb
@@ -37,8 +37,10 @@ module SimpleDeploy
     end
 
     def destroy
-      stack.destroy
-      @logger.info "#{@name} destroyed."
+      if attributes['protection'] != 'on'
+        stack.destroy
+        @logger.info "#{@name} destroyed."
+      end
     end
 
     def events(limit)
